requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
sentence-transformers==2.6.1
faiss-cpu==1.7.4
langchain==0.1.20
numpy==1.26.4
python-multipart==0.0.7
python-docx==0.8.11